\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,DoppelGang-method}
\title{Print a DoppelGang object}
\description{
Prints the @summaryresults slot of a DoppelGang object.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "DoppelGang")}}{
Prints the @summaryresults slot of a DoppelGang object.
}

}}
\keyword{methods}

