\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,DoppelGang-method}
\title{Show a DoppelGang object}
\description{
Formats a summary of the DoppelGang object.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DoppelGang")}}{
Summaryizes the @summaryresults slot of a DoppelGang object
}
}}
\keyword{methods}
