\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,diagonalMatrix-method}
\alias{summary,DoppelGang-method}
\alias{summary,sparseMatrix-method}
\title{ Summarizes a DoppelGang object }
\description{
Summarizes the @summaryresults slot of a DoppelGang object.
}
\section{Methods}{
\describe{
  \item{\code{signature(object = "DoppelGang")}}{
    Summarizes the @summaryresults slot of a DoppelGang object.
  }
}}
\keyword{methods}
